+++ stag_lib/utility.cpp
@@ -42,6 +42,8 @@ std::vector<double> stag::sprsMatToVec(const SprsMat* matrix) {
 }
 
 std::vector<double> stag::sprsMatToVec(const SprsMat* matrix, stag_int n) {
+  if (n < 1) throw std::invalid_argument("Dimension n must be at least 1.");
+
   // Initialise the solution vector.
   std::vector<double> dense_vec;
 
@@ -60,6 +62,17 @@ std::vector<double> stag::sprsMatToVec(const SprsMat* matrix, stag_int n) {
 SprsMat stag::sprsMatFromVectors(std::vector<stag_int>& column_starts,
                                  std::vector<stag_int>& row_indices,
                                  std::vector<double>& values) {
+  // The length of the row_indices and values vectors should be the same
+  if (row_indices.size() != values.size()) {
+    throw std::invalid_argument("Sparse matrix indices and values array length mismatch.");
+  }
+
+  // The last value in the column_starts vector should be equal to the length
+  // of the data vectors.
+  if (column_starts.back() != row_indices.size()) {
+    throw std::invalid_argument("Final column starts entry should equal size of data vectors.");
+  }
+
   SprsMat constructed_mat = Eigen::Map<SprsMat>((stag_int) column_starts.size() - 1,
                                                 (stag_int) column_starts.size() - 1,
                                                 (stag_int) values.size(),